# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Application API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the generated application",
    "contact": {
      "name": "API Support"
    }
  },
  "servers": [
    {
      "url": "https://your-production-url.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
  "security": [],
  "paths": {
...
    "/api/leads": {
      "get": {
        "summary": "Get all leads with filters",
        "tags": [
          "Leads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "assigned_to",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Leads retrieved successfully"
          }
        }
      },
      "post": {
        "summary": "Create a new lead",
        "tags": [
          "Leads"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Lead created successfully"
          }
        }
      }
    },
...
    "/api/sales/{id}": {
      "get": {
        "summary": "Get sale details",
        "tags": [
          "Sales"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sale details retrieved"
          }
        }
      },
      "put": {
        "summary": "Update sale status",
        "tags": [
          "Sales"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sale updated successfully"
          }
        }
      }
    },
    "/api/sales": {
      "get": {
        "summary": "Get all sales with filters",
        "tags": [
          "Sales"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sales retrieved successfully"
          }
        }
      },
      "post": {
        "summary": "Create sales transaction",
        "tags": [
          "Sales"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Sale created successfully"
          }
        }
      }
    },
    "/api/service/appointments/{id}": {
      "put": {
        "summary": "Update appointment",
        "tags": [
          "Service"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Appointment updated successfully"
          }
        }
      }
    },
...
  },
...
}